Fix implicit conversion of returned value in Touch::numTouches()

+++ framework/include/minko/input/Touch.hpp
@@ -81,7 +81,7 @@ namespace minko
             int
             numTouches()
             {
-                return _identifiers.size();
+                return static_cast<int>(_identifiers.size());
             }
 
             inline
